requirements.txt
einops
transformers
omegaconf
tqdm
pandas
numpy
opencv-python
pillow
scikit-image
accelerate
diffusers
wandb
timm
decord
seaborn
scikit-learn
prettytable
pytorch_lightning
lightning
beartype
piq
rotary_embedding_torch
